requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0
chromadb==0.4.17
pdfplumber==0.9.0
python-multipart==0.0.6
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3